TD-6: add index cfg for unit tests
