+++ tests/unit/store/modules/taskStore.cfg.js
